chore: set up first projects
Set up the initial projects for a simple project, complex project,
TypeORM project, and a MongoDB project.
The simple project will be nothing more than a controller, a service,
and a pipe.
The complex project will be a few controllers, a few services,
some interceptors, pipes, a filter or two, and maybe authentication.
The TypeORM project will be focused on TypeORM usage and testing.
The MongoDB project will be focused on MongoDB usage and testing.