Dockerize (#601)
* chore: moves to eslint and starts dockerization

* fix(websocket): fix jest not exiting properly

Because I was using `io.connect()` without a properport, socketIo was not properly disconnecting
after the test. Now that I am useing `io.connect(await app.getUrl())` everything works just fine

fix #14

* ci: starts dockerizing the e2e testing process

I'm tired of seeing CRON failures so I'm also adding in the necessary service to the cron.yml for
the moment. I'll get it all fixed in the future

* test(typeorm): fixes up typeorm e2e tests with docker-compose

* feat: removes typeorm-graphql-sample

The TypeOrm-GraphQL-Sample project is removed as there is no need for two sample projects to use
TypeORM. The point of the repository is to show how testing can be done with NestJS. As there is a
fully working typeorm-sample project and a graphql-sample project, there is no need to combine the
two.

BREAKING CHANGE: removal of the typeorm-graphql-sample

* feat(e2e): adds docker-compose for e2e tests

There's still some bugs with how the microservices tests were implemented, I'll need to spend more
time on that, but now the TypeORM test has an actual database to query and use. Mongo is next, but I
want to get rid of the failing CRON test for now.

* fix: fixes microservice tests