chore: migrate to nx monorepo
I'm moving over from a Nest monorepo to an [nx][1] monorepo to allow for
easier development and more robust commands. The ability to use `npm run nx affected:test`
to take care of all of the changed tests at once is incredibly useful. Also, it helps with
running the projects locally via compiling with webpack, and generally has a better DX in
my honest opinion. Plus, it can allow for a useful refactor pattern later, and better exposure
to using Nx.

closes #1222

[1]: https://nx.dev